+++ genvalidity-hspec-optics/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## [Unreleased]
 
+## [0.1.1.0] - 2018-10-06
+
+### Added
+
+* `shrinkUncheckedDoesNotShrinkToItself`
+* `shrinkUncheckedDoesNotShrinkToItselfWithLimit`
+* `shrinkValidDoesNotShrinkToItself`
+* `shrinkValidDoesNotShrinkToItselfWithLimit`
+* `shrinkInvalidDoesNotShrinkToItself`
+* `shrinkInvalidDoesNotShrinkToItselfWithLimit`
+
+### Changed
+
+* Added the concept that shrinkValid should not shrink to itself to `shrinkValidSpec` and `shrinkValiditySpec`.
+
 ## [0.1.0.0] - 2018-08-25
 
 ### Added